requirements.txt
discord==1.0.1
NBT==1.5.0
Pillow==7.2.0
requests==2.24.0

